(SERVER-2172) Pull out spec util methods
